GPS좌표와 구글맵 연동 완료.
* 사용자 이동에 따라 맵의 변화를 확인함.

* 줌인/줌아웃 기능이 필요해보인다.

+++ Traveling/app/src/main/java/com/example/edwin/traveling/System/LocationService.java
@@ -32,8 +32,8 @@ public class LocationService extends Service {
                 presentLocation.setLongitude(location.getLongitude());
 
                 Intent intent = new Intent("PosData");
-                intent.putExtra("LA", location.getLatitude());
-                intent.putExtra("LO", location.getLongitude());
+                intent.putExtra("LA", (float) location.getLatitude());
+                intent.putExtra("LO", (float) location.getLongitude());
                 getApplicationContext().sendBroadcast(intent);
 
             }
